move tar and --bypass-lock tests to own modules
